LATX, fix: Fixed the error of pir1_base in get_ir1_list()
Fixed app: winecfg for wine-10.9-x86.

Reason for the problem: The pir1_base used in get_ir1_ist() is the base
address of the current TU, not the base address of the current TB. The
pir1_base used when handling disassembly failure is incorrect.

Fixed method: When executing the get_ir1_ist() function, set pir1_base to
the current TB's pir1_base.