LATX, fix: smc opt support more inst
The SMC code-first optimization is enabled by default. But not all store inst
are supported to be interpreted after smc.

This patch supports some more store inst (stptr, fst, vxst) to be interpreted
in smc strategy. To avoid abort, when the store is unsupport or sc.w/d/q,
the strategy on this page will fallback to the original write-first way,
which means all TB on this page will be invalidated to make this page writable
and return back to execute the original store inst.

A simple profiling work is performed for this fallback case.  If this case
occurs too much, some info will be printed to indicate futher optimization.

Signed-off-by: NiuGenen <niugenen@qq.com>