LATX, fix: spm set and clear
The spm (shadow page map) is an optimization for the interval tree of the
ShadowPageDesc to avoid expensive iteration. If the spm_exist return
false, then there is no need to lookup the interval tree. This patch fix
the set and clear of the spm.

This patch fix the WXWork (企业微信) coredump problem. The frontend
WXWork works fine, but the backend process causes a lot of coredump.
They are frequently crashed and recreated again, which consums a lot of
CPU resource.

Signed-off-by: NiuGenen <niugenen@qq.com>