LATX, fix: tb reset offset when lazylink
In tblink optimization, the link stub (b + nop) is generated inside TB
and will be rewrite to implement direct jmp from one TB to another.

There are two key variable of TB that are related to this tblink optimization.
- jmp_target_arg[n] points to the start of the link stub
- jmp_reset_offset[n] points to the end of the link stub, which is also
  the next inst after the link stub

These two variables are calcualted in the label_dispose().
And the size of the link stub is fixed (B_STUB_SIZE).

But in the lazylink optimization, the link stub is eliminated. The
B_STUB_SIZE should be zero. Otherwise the jmp_reset_offset[n] might
points to an address that locates outsize of this TB.

Signed-off-by: NiuGenen <niugenen@qq.com>