LATX, fix: lock interrupt generate segv
Some atomic instructions are translated with ll.d/sc.d to maintain
atomicity.  If the ll.d does an unaligned access to a page that has
PAGE_READ but not PAGE_WRITE, the SIGBUT will be trigered first.
However, the lock_interpret() does not check wether the page has
PAGE_WRITE, or PAGE_WRITE_ORG more specifically.

This patch adds the checking of PAGE_WRITE_ORG in lock_interpret().
To generate correct SEGV for guest, the siginfo is modifyed and the
pc in uc context is changed to point to a store inst.

Signed-off-by: NiuGenen <niugenen@qq.com>